Add Footer and ContactForm components with responsive design and form handling
